heterodyne: Introduce REF_SPECTRUM_FILE recipe parameter.
This gives a final route to correct for reference (off-position)
spectra signal through a user-derived NDF.  At present there needs to
be some valid REF_SPECTRUM_REGIONS extents supplied too for the NDF to
be subtraccted from the spectra, but I plan to change that. If just
the NDF subtraction is required, a range whose limits lie within the
extent of a line spectfied in the NDF could be supplied.