Calib: Validate receptor names by instrument in list system.
The --calib bad_receptors= command-line syntax permits additional
receptors to be flagged as bad, in the so-called list system.
ORAC::Calib::ACSIS tested that the supplied receptors were valid, but
was only comparing with the HARP set H00..H15.  I've added a new
method, receptor_names, which checks against appropriate receptors
names for the various current, historical, and upcoming Namakanui
instruments.  If a invalid receptor is supplied, it now issues a
warning rather than silently continuing leaving the ORAC-DR user
puzzled or none the wiser.

We should look for other places where the receptor name is assumed to
be H00..H15.