heterodyne: Extract the emission-finding from _CREATE_MOMENTS_MAPS_THROUGH_SMOOTHING_.
Create a new primitive to locate emission lines.  The goals are to
permit its use in more than one place, to be easier to test and improve,
and to make the 1000-line _CREATE_MOMENTS_MAPS_THROUGH_SMOOTHING_ more
manageable.  Remo Tilanus wanted this many years ago, as documented in
his comments within _CREATE_MOMENTS_MAPS_THROUGH_SMOOTHING_.  The
obstacle seemed to be handling the $colin temporary NDF.  To overcome
this the calling primitive specifies the name of non-temporary NDF with
the default "_emis" suffix.  This NDF can be added to the intermediate
files to remove it at the recipe's completion.

In addition to the prologue, argument handling, and indentation
adjustement, some of the comments were augmented, improved, or
corrected.  The temporary file names were renumbered as $colin (now
$mask via Argument MASK) was from the first such file ($temp1 in the
original source).  The primitive also returns the number of emission
lines detected.