heterodyne: Simplify the smoothing code.
The ability to smooth a cube has been around for about 12 years, and I
strongly doubt anyone is running a version of KAPPA from that time.
Therefore, it seems appropriate to remove the test to perform
smoothing in frequency and spatially in two passes.  This all helps
shortening the code, especially in _CREATE_MOMENTS_MAPS_THROUGH_SMOOTHING_.