heterodyne: Introduce CUBE argument.
This is to avoid copying a Group object into a Frame object within the
calling primitive in order to pass the input cube name.  The name is
passed via new Argument CUBE.  Also an undefined Argument FILE
defaults to the value of CUBE.

The primitive still uses the current Frame to access metadata.