SCUBA2: return COSYS from _GET_MAP_CENTRE_
The coordinate system is returned if this primitive constructs
new coordinates from BASEC1 and BASEC2.  _DEFINE_APERTURE_ can then
use it instead the value it obtains from wcsattrib.

(They may not be the same, for example, if the map has been made
in a coordinate system other than that in which it was observed.
However this does not solve the mismatch for pointings, which
use AZEL offset coordinates.)