lib: Always include results in screen output
Previously results were written to stdout with "--log s" but not
with "--log sf".  (This was because "s" leaves settings at their
defaults, which included stdout for results, whereas anything
else leads to explicit configuration, where results were excluded
from screen output.)  Including the results in the explicit
configuration for screen output makes these two modes consistent.
It is also useful to be able to see the results messages without
having to use an X window.