heterodyne: Have a file suffix to discriminate masked and unmasked products.
This is during the subtraction of the reference spectrum.  The suffix
applies to both the _ressp (reference source-subtracted spectrum) and
_mrsp (mean refereence spectrum) products.  The suffix is m for masked
and um for unmasked.  This helps diagnosis when the filtering has not
worked as expected.

This version compiles and works, creating the differently named files
while the resultant spectral cube is unchanged.  In addition to
correcting the misplaced paste in Commit 98faca00b8f56, I made
defensive-programming additions for the suffix naming, and passed the
correct suffix (i.e. includes the trailing m or um) to _ALIGN_SPECTRA_.