heterodyne: Document ALIGN_SIDE_BAND recipe parameter.
ALIGN_SIDE_BAND is listed in SC/20 Appendix G, and is active in the
four recipes mentioned there, but it was not documented in the recipe
prologues..  That omission is corrected here.

Also report ALIGN_SIDE_BAND, if set, regardless of whether it is
disabled or enabled.  This had been done for the narrowline recipe,
but not the other three recipes.