heterodyne: Untag "emis" file from _FIND_EMISSION_
When _FIND_EMISSION_ uses a file which has been tagged as representative
or the reduced product, its output can retain the corresponding FITS
headers.  This commit adds a new primitive _UNTAG_FILE_ to remove
the PRODUCT and PRODID files so that the "emis" file is not considered
a product by jsawrapdr.  (An "emis" product causes JSA processing
jobs to fail due to the presence of a product with an unexpected name.)
Possibly better approaches might instead be to not use a tagged file as
input for further processing, or to ensure the "emis" file is deleted
at the end.