lib: Add SCUBA-2-style number pattern for ACSIS
This updates the ORAC::Frame::ACSIS::number method so that the first
pattern tried matches from the start of the file name, i.e. with
aYYYYMMDD_(number)_..., similarly to the first pattern for SCUBA-2.
The original pattern, which matches from the end of the name, is
retained (as the last tried) in case the new pattern would fail in
a situation where the original pattern would have worked.  The new
pattern should allow matching against files processed by Wesley, to
which a suffix has been appended.