heterodyne: Group files by time for hybrid merge
This is to avoid errors which occur when trying to merge files
which do not overlap in time, i.e. when ACSIS wrote out the raw
data as multiple subscans.

Note: an alternative, probably better, solution would be to merge
the smurf-8byte branch of Starlink and alter the timesort step
to combine each subsystem into a single file.