lib: Add option to read flag files in sync
This is controlled by a new command line option "flagsync" so that
the original default behavior is retained.  If enabled, only matching
numbers of entries will be read from each of a group of flag files
being read at the same time.  This is intended for instruments
which write multiple flag files, such as SCUBA-2, so that the summit
pipeline receives consistent numbers of subscans for all subarrays.