wesley: Add recipe to make OBJECT hdr. consistent
When a group would contain inconsistent values of the OBJECT
header, this recipe can be used to set all subsequent values
equal to the first found.  This uses a new primitive
_GET_GROUP_HEADER_ to store the first value found as an
Astro::FITS::Header::Item instance in the group uhdr.
Also _SET_HEADERS_ was updated to accept Astro::FITS::Header::Item
objects and to attempt to format them suitably for the
fitsmod file.