wesley: Ensure output files are not in $Frm->raw
This is to prevent the pipeline deleting them at the end of
processing if they are symlinks.  (An alternative would be
for _WESLEY_GOODBYE_ to recognize that an output file is still
a symlink and create a new symlink to the same file with a
suffix.  However retaining the original file name may make it
easier for archiving software to detect unchanged raw files.
If this change is retained as-is, it would also be possible,
for example, for _SET_HEADERS_ not to copy files for which
no changes are required.)