bin: Overlook ORAC_DATA_IN check if --files abs.
If the "--files" option has been used to specify a list of absolute
pathnames, do not exit with an error because ORAC_DATA_IN
is undefined or does not refer to a directory which exists.
Instead in this case define a dummy value for ORAC_DATA_IN
in case anything still tries to access it.

It might be possible to provide a better default for ORAC_DATA_IN,
e.g. the parent directory of the specified file list.  (This would
allow the file list to contain paths relative to its own location.)
However this commit makes a minimal change to the behavior of
ORAC-DR just for the specific case that a file list consisting
entirely of absolute pathnames has been given.