lib: Add HLRANGE option for KAPVIEW graph
This is to allow a section of the graph to be highlighted.  Entries for
the style parameter are now accumulated in an array so that the
parameter can be constructed at the end.  The clear parameter is
now given directly in the linplot/mlinplot command rather than
being part of the $args string to allow it to be omitted when
highlighting a section.