heterodyne: Highlight collapse part of p/f spectra
When doing a multiline graph of pointing / focus spectra, attempt
to highlight the section which was collapsed for the line results.
_COLLAPSE_CENTRAL_LINE_ and _COLLAPSE_LINE_ set a uhdr COLLAPSED_RANGE
indicating the range used.  Displaying the focus cube is now done
in _REDUCE_LINE_FOCUS_ so that this uhdr is available.

Unfortunately we don't seem to be able to use a WCS-based NDF
section of a file which ndfcopy has trimmed.  Therefore for now,
_COLLAPSE_LINE_ converts its WCS section to a pixel range.