heterodyne: Check type of pointing catalog entries
When looking for the current source in the pointing catalog to
get the velocity range over which to collapse, check the type
given at the start of the comment.  We need to exclude continuum
entries (used with SCUBA-2) with type [c] since these generally
do not specify the range.  Also with the addition of Alaihi with
its own [w]-type pointing sources we should distinguish these
from [l] if we recognize the molecule (SiO this case).