RxH3: implement some frame flag file methods
This is to allow more looping options to be used with RxH3.  Since
these flag files (and the associated raw data files) are named
by date and time, we need to check the OBSNUM FITS header in the
data associated with each flag file.  A dummy file name is returned
if there is no flag file for the requested observation.

Since we do not generally have a large number of RxH3 observations
per night, reading the FITS headers should not be too problematic.
Additionally the observation number associated with each flag file
is cached in the ORAC::Frame::RxH3 module so each need only be
checked once.