SCUBA2: make running obs mosaic for summit preview
For long observations, a preview was previously produced for
each section processed, overwriting the previous preview image.
Now _CREATE_PREVIEW_SUMMIT_ can use _MAKE_MOSAIC_ to create
a running co-add, as for the group mosaic, but starting afresh
each observation.  To keep the preview generation code separate
from the main operation of the pipeline, _MAKE_MOSAIC_ is called
with a temporary group object, and the running co-add is tracked
by a new group uhdr RUNNING_COADD which is cleared at the start
of an observation by _SUMMIT_STEER_.