heterodyne: Validate the integ map.
If the emission spans the entire integ, do not use it during
standing-wave removal.  Note that this does not return a GOOD argument,
but resets the STANDING_WAVE_EMISSION_MASK recipe parameter that will be
accessed in the standing-wave subtraction primitives, accessed later.

Record the standing-wave corrected array by reference in the Group
header STANDING_WAVE_CORRECTED.