heterodyne: Avoid leaving _varNNN files behind
When $group == 2, we can not use $Obj->push_intermediates
because $Obj might be a frame for which _DELETE_TEMP_FILES_
has already been run.  Therefore instead push to the
$Grp intermediates in this case.  This was particularly
problematic if the _varNNN files were based on a previous
reduced product and so had headers (e.g. PRODUCT) causing
jsawrapdr to attempt to store them in archive processing.