heterodyne: Preserve reference pixel for raster
Commit 34721733ee introduced Frame uhdr values REFPIX1 and REFPIX2
to record the reference pixel values used with makecube, but only
when the observation type was not "raster".  As a result some
intermediate products could be tiled differently, potentially
leading to an error where a "cubemask" file has no pixels in common
with the corresponding tile.  Therefore add a second point
where these uhdr can be set for type "raster".  Also apply
the same rounding to the original uhdr values as will be
used in the "raster" case.