PICARD: adjust MAKE_HIPS recipe
Two small primitives _MAKE_HIPS_PROPERTIES_ and _MAKE_HIPS_ALLSKY_
have been extracted from the main _MAKE_HIPS_ primitive.  The
_EXTRACT_JSA_TILES_ step has been moved from the recipe to the
main primitive so that it can be made optional based on a
new recipe parameter INPUT_JSA_TILES.

One intermediate file step was removed at the initial tile splitting
step to save a little time here.  (The recipe as a whole is still
extremely slow.)

For the "allsky" file, the tiling was corrected (27x29 instead of
27x28 at order 3) and the background set to bad rather than value 0.

To allow the frame processing to perform incremental clean-up,
the _DELETE_TEMP_FILES_ primitive was updated to add a new
parameter RETAINKEPT.  When this is set, it retains kept
files in the frame/group's intermediates list so that they
can be reconsidered for deletion again later.