PICARD: generate HiPS structure recursively
Alter the _MAKE_HIPS_ primitive so that instead of working through the
HiPS orders one at a time, it instead assembles a hierarchical data
structure which records how the input files will be combined to make
the tiles at each order.  This is then passed to _MAKE_HIPS_TILE_
which operates recursively to create each tile by preparing its
constituent parts.  (Note that since primitives can't actually
be recursive, it operates via an internal recursive subroutine.)
Operating in this manner allows temporary files to be deleted
as soon as they are no longer needed, thereby reducing the amount
of disk space required in the output directory.