RxH3: add recpar OFFSET_UNITS
This can be used to request "physical" offsets, in which case a
geometrical factor is applied to convert distance along the beam to a
measurement normal to the surface.  Note that, instead of giving the
parameter "physical" to holophasedist, this conversion is performed
using kappa maths, after fitting and subtracting deformations.
This is to allow for more accurate fitting where the deformation
model (e.g. defocus) is in terms of distance along the beam.
However the NDF label is updated to "Offset (physical)" (as would
have been used by holophasedist) so that fitpanels recognizes that
this geometrical factor has already been applied.  The requested unit
type is added to the log.surface file.