bin: Add --input-file option to oracdr
This adds an input file to the list of files which may have been
read using the --files option.  It can be useful when processing
a single file.  (A few files can also be specified by repeating
the option.)  The option name was chosen to avoid conflicting
with (possibly abbreviated forms of) the --files option.  It can
of course itself be abbreviated, e.g. to --input.