heterodyne: Restrict number of spectra in pntcube
Since this is to be displayed as a "multiline" plot (with mlinplot)
it should not contain an excessive number of spectra.  If the dimensions
are more than 3x3 (as for a 5-point) then try to extract the section
from 0:2 on each axis, if this overlaps the bounds, or the central
pixel otherwise.  Use permaxes rather than trimming to ensure the
2nd axis is not discarded if only one spectrum is selected.