heterodyne: Grid HARP pointings in tracking system
HARP pointings are typically performed in tracking coordinates
(identified by the FITS header JIG_CRD = "FPLANE").  We must
grid in these coordinates in order to be able to fit the pointing
offsets correctly.  Added a new primitive "_ROTATE_TO_AZEL_"
which rotates the image to be upright in the Azimuth-Elevation
system for the purpose of the real-time display and preview
image.  This is so that the display can still serve as a useful
guide to the direction of the pointing offset.  When the observation
system is not Azimuth-Elevation, the normal display of the
collapsed line and continuum maps is suppressed via a new
DISPLAY parameter to the relevant primitives.