heterodyne: Pass system to _GET_MAKECUBE_PARAMS_
If a coordinate system has been specifically requested, pass it
to the _GET_MAKECUBE_PARAMS_ primitive.  This is so that this
primitive can take the choice into account.  (Although currently
_CREATE_CUBE_FRAME_ and _CREATE_CUBE_GROUP_ both override the
system with that requested again later.)  This should prevent
warnings that JIG_CRD is undefined for focus observations since
_CREATE_FOCUS_CUBE_ currently calls _CREATE_CUBE_FRAME_ with
SYSTEM=AZEL.