heterodyne: Add TRIM arg. to collapse primitives
This controls the "trim" parameter passed to the kappa collapse task.

Enable this option in _REDUCE_CONTINUUM_FOCUS_ and _REDUCE_LINE_FOCUS_
so that _CALCULATE_FOCUS_ receives a trimmed file and does not need
to trim (with ndfcopy) as its first operation.

Note: removing the spectral axis causes problems for header syncing
when the ACSIS frame attempts to use the WCS to obtain the rest
frequency.  To avoid this, set a RESTFREQ header before trimming since
the "rest_frequency" method will check this first.