SCUBA2: add opt. to format log entry params/values
This adds to the _ADD_LOG_ENTRY_ primitive optional arguments
ENTRY_FORMAT and PARAM_FORMAT which can be used to give
sprintf-style formatting strings for the values and titles of
the log files.  If these are present, they are used to format the
values, titles and units, after enclosing the units in brackets.
If not given, these inputs are assumed to already have been
formatted (and the units enclosed in brackets) as before.