SCUBA2: make REDUCE_POINTING more like REDUCE_SCAN
This updates the REDUCE_POINTING recipe to operate more like
the standard REDUCE_SCAN recipe.  The map is made in the
default coordinate system (rather than Az-El).

The step of tagging a group reduced product has been removed
since ORAC-DR puts pointings in groups of a single observation
each, so a group product may well not be useful.  As the recipe
stands, at the point where a group product could be tagged,
the map will have been cropped, so the recipe would need to be
altered to crop later if it were necessary to generate
a group reduced product.

A CALONLY argument was added to _CALCULATE_FCFS_ to ensure that
the cropping step is applied to all pointing sources. (The name
of the argument was chosen to match that in
_REMOVE_BACKGROUND_FROM_IMAGE_.)