SCUBA2: change suffix of _FIND_VARIANCE_ trim file
Updated the _TRIM_MAP_EDGES_ primitive with "SUFFIX" and "SETTAG"
arguments so that, when _FIND_VARIANCE_ calls it, it can use names
other than "_crop" and "crop".  This is to avoid overwriting
the cropped file generated by _CROP_IMAGE_ALL_ (e.g. inside
_CALCULATE_FCFS_).  REDUCE_POINTING can then find pointing offsets
from the intended square cut out rather than the temporary cropped
file made to calculate the variance.  The argument name "SETTAG"
was used because some SCUBA-2 primitives use "TAG" to specify a tag
from which to obtain files on which to operate.