general: Extract prim. to prepare point. log info
This extracts some of the logic in the SCUBA-2 and heterodyne
pointing primitives which prepares information to include in
pointing log files.  It also adds new parameters "URA" and "UDEC"
which give the (estimated) UAZ and UEL values (after this pointing)
rotated to RA, Dec coordinates (as at the start of the observation).