SCUBA2: update _DISPLAY_ZPD_ to handle scan dir.
If "FTS_SCANDIR_ZPD" is enabled, organize the files by scan
direction before mosaicking for display, and add "fwd" or "bck"
to the filename suffix to distinguish the two types.  In either
case avoid using "ALLOW_NUMBERED_SUFFICES" (via a temporary frame)
so that numbers are not included in the suffixes and they match
the entries in disp.dat.  (New "fwd" and "bck" variants have been
added to disp.dat, with those in the "bck" direction shown in a
different window.)