SCUBA2: swap inversion for 850um subarrays
In commit e55329cfdc the ZPD interferogram inversion criterion (based
on which shutter is in the beam) was swapped for the 450um subarrays.
However in the data taken on 2025-06-25, it looks like the sign of
central peak was the same in the 850um and 450um data.  Therefore
remove the subarray (wavelength) test here and always use the
previous 450um criterion (invert if shutter in this beam)?