heterodyne: Paste im. mask inside numgood if block
In _FLAG_HIGHFREQ_INTERFERENCE_SPECTRA_, move the step of pasting
$temp4file into the image mask $temp7file inside the if block which
starts "if ( $numgood > 0 ) { ...".  This is because $temp4file is
created inside this block.  Applying the paste afterwards causes
an error "Value 2 given is outside the allowed range for parameter
AXES.  Give a value in the range 0 to 2 please" where $numgood is
not > 0 because $temp4file will have been changed to a different
dimensionality using MANIC while processing the previous receptor.
(Or presumably may not even exist if this is the first receptor.)

Note: this error occurred in processing job 387489, at observation 37
- the last in its group - on 2025-08-23, while considering receptor H02.
Receptor H02 is disabled by index.bad_receptors but this primitive
seems to ignore it by doing $Cal->bad_receptors( "INDEX" );